requirements.txt
MySQL-python
djangorestframework>=3.1.3
django-cors-headers
django-rest-auth
sendgrid-django
django-polymorphic==0.7.2
django-extensions==1.5.9
ipdb==0.8.1
python-dateutil==2.4.2

